dnsdist: Move DNSRule declaration to dnsdist-rules.hh
