+++ pdns/dnsdistdist/dnsdist-rules.hh
@@ -36,6 +36,18 @@
 #include "dnsparser.hh"
 #include "dns_random.hh"
 
+class DNSRule
+{
+public:
+  virtual ~DNSRule()
+  {
+  }
+  virtual bool matches(const DNSQuestion* dq) const = 0;
+  virtual string toString() const = 0;
+  mutable stat_t d_matches{0};
+};
+
+
 class MaxQPSIPRule : public DNSRule
 {
 public:
