+++ pdns/dnsdistdist/dnsdist.hh
@@ -895,17 +895,6 @@ void responderThread(std::shared_ptr<DownstreamState> dss);
 
 class DNSDistPacketCache;
 
-class DNSRule
-{
-public:
-  virtual ~DNSRule()
-  {
-  }
-  virtual bool matches(const DNSQuestion* dq) const = 0;
-  virtual string toString() const = 0;
-  mutable stat_t d_matches{0};
-};
-
 struct ServerPool
 {
   ServerPool() :
